# backend/cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/events/add-co-owner": {
            "post": {
                "description": "Tạo sự kiện để thêm một đồng sở hữu mới",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Tạo sự kiện thêm đồng sở hữu",
                "parameters": [
                    {
                        "description": "Thông tin yêu cầu thêm đồng sở hữu",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.CreateAddCoOwnerEventRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.CreateAddCoOwnerEventResponse"
                        }
                    },
                    "400": {
                        "description": "Yêu cầu không hợp lệ",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Lỗi xử lý sự kiện",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/events/burn": {
            "post": {
                "description": "Tạo sự kiện để đốt token với số lượng chỉ định",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Tạo sự kiện đốt token",
                "parameters": [
                    {
                        "description": "Thông tin yêu cầu đốt token",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.CreateBurnTokenEventRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.CreateBurnTokenEventResponse"
                        }
                    },
                    "400": {
                        "description": "Yêu cầu không hợp lệ",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Lỗi xử lý sự kiện",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/events/mint": {
            "post": {
                "description": "Tạo một sự kiện mint token với số lượng chỉ định",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Tạo sự kiện mint token",
                "parameters": [
                    {
                        "description": "Thông tin yêu cầu mint",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.CreateMintEventRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.CreateMintEventResponse"
                        }
                    },
                    "400": {
                        "description": "Yêu cầu không hợp lệ",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Lỗi xử lý sự kiện",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/events/pause": {
            "post": {
                "description": "Tạo sự kiện để tạm dừng hoạt động hợp đồng",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Tạo sự kiện tạm dừng",
                "parameters": [
                    {
                        "description": "Thông tin yêu cầu tạm dừng",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.CreatePauseEventRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.CreatePauseEventResponse"
                        }
                    },
                    "400": {
                        "description": "Yêu cầu không hợp lệ",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Lỗi xử lý sự kiện",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/events/sign": {
            "post": {
                "description": "Ký hoặc từ chối một sự kiện hiện có",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Ký sự kiện",
                "parameters": [
                    {
                        "description": "Thông tin yêu cầu ký",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.SignEventRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.SignEventResponse"
                        }
                    },
                    "400": {
                        "description": "Yêu cầu không hợp lệ",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Lỗi xử lý sự kiện",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/events/unpause": {
            "post": {
                "description": "Tạo sự kiện để bỏ tạm dừng hoạt động hợp đồng",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Tạo sự kiện bỏ tạm dừng",
                "parameters": [
                    {
                        "description": "Thông tin yêu cầu bỏ tạm dừng",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.CreateUnpauseEventRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.CreateUnpauseEventResponse"
                        }
                    },
                    "400": {
                        "description": "Yêu cầu không hợp lệ",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Lỗi xử lý sự kiện",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.CreateAddCoOwnerEventRequest": {
            "type": "object",
            "properties": {
                "jwt": {
                    "type": "string"
                },
                "newCoOwner": {
                    "type": "string"
                }
            }
        },
        "controllers.CreateAddCoOwnerEventResponse": {
            "type": "object",
            "properties": {
                "eventId": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "transactionHash": {
                    "type": "string"
                }
            }
        },
        "controllers.CreateBurnTokenEventRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "jwt": {
                    "type": "string"
                }
            }
        },
        "controllers.CreateBurnTokenEventResponse": {
            "type": "object",
            "properties": {
                "eventId": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "transactionHash": {
                    "type": "string"
                }
            }
        },
        "controllers.CreateMintEventRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "jwt": {
                    "type": "string"
                }
            }
        },
        "controllers.CreateMintEventResponse": {
            "type": "object",
            "properties": {
                "eventId": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "transactionHash": {
                    "type": "string"
                }
            }
        },
        "controllers.CreatePauseEventRequest": {
            "type": "object",
            "properties": {
                "jwt": {
                    "type": "string"
                }
            }
        },
        "controllers.CreatePauseEventResponse": {
            "type": "object",
            "properties": {
                "eventId": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "transactionHash": {
                    "type": "string"
                }
            }
        },
        "controllers.CreateUnpauseEventRequest": {
            "type": "object",
            "properties": {
                "jwt": {
                    "type": "string"
                }
            }
        },
        "controllers.CreateUnpauseEventResponse": {
            "type": "object",
            "properties": {
                "eventId": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "transactionHash": {
                    "type": "string"
                }
            }
        },
        "controllers.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "controllers.SignEventRequest": {
            "type": "object",
            "properties": {
                "approve": {
                    "type": "boolean"
                },
                "eventId": {
                    "type": "string"
                },
                "jwt": {
                    "type": "string"
                }
            }
        },
        "controllers.SignEventResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                },
                "transactionHash": {
                    "type": "string"
                }
            }
        },
...
    }
}